Update poll vote spec to assert broadcast
